Issue #0000 feat: Cleaning up app cache
DO-173

+++ build.sh
@@ -4,13 +4,13 @@ set -o errexit
 e () {
     echo $( echo ${1} | jq ".${2}" | sed 's/\"//g')
 }
-m=$(metadata.sh)
+m=$(./metadata.sh)
 commit_hash=$1
 
-org=$(e "${m}" "org")
+org=sunbird
 name=$(e "${m}" "name")
 version=$(e "${m}" "version")
 
-git archive --format=tar.gz sunbirdntp inventory --output=code.tar.gz
+[[ -f code.tar.gz ]] && rm -rf code.tar.gz
+git archive --format=tar.gz ${commit_hash} --output=code.tar.gz 
 docker build -f ./Dockerfile --label commitHash=${commit_hash} -t ${org}/${name}:${version}-bronze .
-rm -rf code.tar.gz
